token 재발급시 중복 요청 이슈 해결. 그런데 재발급 속도가 너무 느림

+++ src/main/resources/static/js/api.js
@@ -57,6 +57,18 @@ const fetchWithAuth = async (url, options = {}) => {
                     method: 'POST'
                 });
 
+                if(reissueResponse.status === 429){
+                    await addRefreshSubscriber((newToken) => {
+                        const retryHeader = {
+                            ...options.headers,
+                            'Content-Type': 'application/json',
+                            'access': `${newToken}`
+                        }
+                        resolve(fetch(url, {...options, headers: retryHeader}));
+                    });
+                    return;
+                }
+
                 if(!reissueResponse.ok){
                     console.error('Failed to reissue access token, redirecting to login');
                     sessionStorage.clear();
